+++ client/blocks/authentication/style.scss
